:sparkles:升级为16位数的任意打印

+++ utils.c
@@ -1,10 +1,11 @@
 #include "utils.h"
 #include <stdio.h>
+#include <limits.h>
 
 #define bin2dec_p(...) _bin2dec_p(PP_NARG(__VA_ARGS__), __VA_ARGS__)
 #define dec2bin_p(...) _dec2bin_p(PP_NARG(__VA_ARGS__), __VA_ARGS__)
 
-unsigned dec2bin(unsigned num) {
+unsigned long long dec2bin(unsigned long long num) {
   if (num == 0) {
     return 0;
   } else if (num == 1) {
@@ -13,7 +14,7 @@ unsigned dec2bin(unsigned num) {
   return (num % 2) + 10 * dec2bin(num / 2);
 }
 
-unsigned bin2dec(unsigned num) {
+unsigned long long bin2dec(unsigned long long num) {
   if (num == 0) {
     return 0;
   } else if (num == 1) {
@@ -26,7 +27,7 @@ void _dec2bin_p(size_t argc, ...) {
   va_list ap;
   va_start(ap, argc);
   for (int i = 0; i < argc; i++) {
-    printf("%08u\n", dec2bin(va_arg(ap, unsigned)));
+    printf("%016llu\n", dec2bin(va_arg(ap, unsigned long long)));
   }
   va_end(ap);
 }
@@ -35,7 +36,7 @@ void _bin2dec_p(size_t argc, ...) {
   va_list ap;
   va_start(ap, argc);
   for (int i = 0; i < argc; i++) {
-    printf("%08u\n", bin2dec(va_arg(ap, unsigned)));
+    printf("%016llu\n", bin2dec(va_arg(ap, unsigned long long)));
   }
   va_end(ap);
 }
